Move query to WorkweekHustle
